Add files via upload
Project name- "TITANIC SURVIVAL PREDICTION"

Overview
This project aims to predict the survival of passengers onboard the Titanic using machine learning techniques. The sinking of the RMS Titanic is one of the most infamous shipwrecks in history. On April 15, 1912, during her maiden voyage, the Titanic sank after colliding with an iceberg, resulting in the loss of many lives. This project utilizes the Titanic dataset, which contains information about passengers such as age, gender, class, and whether they survived or not.